Use the GLOW_PATH environment variable to search for apps.
If GLOW_PATH is not set, default to the dapps directory in the repo.

This patch also removes the mukn/glow/ prefix from apps, as it seems
redundant (it's just stripped off immediately), and a bit weird if the
user is going to be able to supply their own apps, not necessarily
written by mukn...

This should probably be mentioned in reference documentation somewhere;
I'm not sure if there's already an appropriate place to document the
behavior of app search?

Another thing to note is that this has the effect of listing all of the
.glow files in dapps/, since the list is no longer hard coded.

A possible enhancement would be to enumerate apps recursively; right
now the prompt just lists .glow files in the top level search
directories.