Remove unnecessary & harmful &check-timeout!.
As-is, if the timeout occurs then the contract is dead -- any attempt to
call it will revert. But we want it to distribute the escrow, so that's
not acceptable. The compiler generates calls to
'check-participant-or-timeout!' when it sees a 'set-participant', so
this isn't strictly necessary anyway.

Note however, that (iiuc) removing this means that a participant can
still send in a message "late" as long as they do so before the other
participant invokes the contract to claim the escrow.