Be better about cleaning up subprocesses.
The escrow timeout integration test now cleans up after itself, so
subsequent tests pass, as do subsequent runs of all tests.

Previously, we were both failing to wait for glow, and glow was failing
to wait for the handshake command. Due to the latter, subsequent runs
could actually get agreement mismatch errors, because one side would end
up talking to the wrong netcat process.

Just waiting for nc is insufficient; if no connection was established,
it will not exist when we close its stdio, so we have to explicitly kill
it.