Incorporate small functions into brk-start.
Now that this is in place, we need one more thing when computing scratch
space: global variables.

This isn't super well tested; I've put in some debugging prints to make
sure the changes in values seem reasonable, but that's it.  Testing this
is somewhat challenging, since so far we don't do dynamic allocation, so
even if brk-start is wrong nothing will actually break.