Track balances for all assets.
...not just the native token.

Along the way, this changes the block context so that deposits &
withdrawals are keyed on asset symbols, rather than the asset
objects themselves.

This breaks the buy_sig integration test; need to investigate more but
don't want to leave too much uncommitted. The failure is that the CALL
instruction that does the withdraw is failing. For some reason, it's
trying to send a value of zero.